+++ packages/analytics/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @gomcp/analytics
+# @mcploom/analytics
 
 ## 0.2.0